Fix #2 - Add ability to enable and disable the system Spell Checker (#42)
Add ability to enable and disable the system Spell Checker

This adds the ability to toggle the spell checker setting for tests. This is useful as sometimes the spell checker dialog can appear during tests and cause tests to fail.

No tests were added for this feature in this commit. Testing the spell checker is difficult to automate -- Android allows you to provide a replacement for the system spell checker, but it must be manually enabled by the user. Adding code to allow this in tests would add more code that could break, making the test more dangerous than productive.