Add keyboard disabler (#43)
* Add keyboard disabler

This adds a new feature to the TestButler API that allows users to tell the system to prefer the hardware keyboard during tests. Often when running UI tests, entering text can cause the software keyboard to rise up. This is a major source of flakiness, so TestButler allows you to prefer the hardware keyboard, which will show nothing on screen.