Calling out that separate APK is required.
I was very confused on how this worked.  I didn't know how it was working, but assumed that the APK was auto-installed after I added the `androidTestCompile` gradle dependency and the necessary lines in my custom `AndroidJUnitRunner`.

The goal of this change is to point out that the separate APK is needed on the device to get Test Butler to Work.