Fixing lint warnings on master (#98)
* Fixing lint warnings on master

Travis was not running for a while, and some lint warnings made it in. All of these warnings/errors are red herrings however:

- Added "PrivateApi" back to the disabled list for the test-butler-app-core module. The whole point of test butler is to invoke private apis.
- Added "WakelockTimeout" back to the disabled list for the test-butler-app-core module. Test butler intentionally grabs an untimed wakelock to keep the device on
- Ignored a specific use of Build.SERIAL. Usually this is not recommended to be used in favor of user controlled ids like AdvertiserID, but this is needed by the infra to uniquely identify devices when talking over adb for real device test butler. We need a guaranteed unique id.

* Fix javadoc generation

Javadoc cant find the aidl class because it doesn't link against it. We can just not link against the class in the javadoc format to avoid the issues while keeping the explanations.

* bump travis config to 29